Add reader-test

+++ lib/Epub/Epub/BookMetadataCache.h
@@ -41,7 +41,7 @@ class BookMetadataCache {
 
  private:
   std::string cachePath;
-  size_t lutOffset;
+  uint32_t lutOffset;
   uint16_t spineCount;
   uint16_t tocCount;
   bool loaded;
